# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_deepakgudla_bookvault_internal_dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_deepakgudla_bookvault_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
...
  github_com_deepakgudla_bookvault_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_server.GraphQLResponse:
    properties:
      data: {}
      errors: {}
    type: object
host: localhost:1357
info:
  contact:
    email: deepakgudla35@protonmail.com
    name: Deepak Gudla
    url: https://www.linkedin.com/in/deepakgudla
  description: A modern e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bookvault-API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the refresh token and log out the user
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: User logout
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request data or user already exists
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/carts/:
    get:
      description: Retrieve current user's shopping cart with all items
      produces:
      - application/json
      responses:
        "200":
          description: successfully retrieved cart
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: cart not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
  /api/v1/carts/items:
    post:
      description: users can add products to their cart
      parameters:
      - description: item added to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully added products to the cart
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "400":
          description: invalid request data or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Add items to cart
      tags:
      - Cart
  /api/v1/carts/items/{id}:
    delete:
      description: users can delete cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted item from the cart
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid cart item
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete items from cart
      tags:
      - Cart
    put:
      description: update the quantity of an item in the user's cart
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: New quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: cart item updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CartResponse'
              type: object
        "400":
          description: invalid request data or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update items to cart
      tags:
      - Cart
  /api/v1/categories:
    get:
      description: retrieve all categories that are active
      produces:
      - application/json
      responses:
        "200":
          description: categories fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get All Categories
      tags:
      - Categories
  /api/v1/categories/:
    post:
      consumes:
      - application/json
      description: create a new product category (Admin Only)
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: successfully created category
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /api/v1/categories/{id}:
    delete:
      description: delete a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: successfully deleted category
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid category ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated category
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CategoryResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
  /api/v1/orders/:
    get:
      description: get the list of orders by user
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's orders
      tags:
      - Orders
    post:
      description: create an order from the current user's cart
      produces:
      - application/json
      responses:
        "201":
          description: successfully created order
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
              type: object
        "400":
          description: empty cart or insufficient stock
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create an Order
      tags:
      - Orders
  /api/v1/orders/{id}:
    get:
      description: get the information about the specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: order retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: order not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
  /api/v1/products:
    get:
      description: get the list of active products
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched products
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
                  type: array
              type: object
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get all products
      tags:
      - Products
  /api/v1/products/:
    post:
      consumes:
      - application/json
      description: create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: successfully created product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Categories
  /api/v1/products/{id}:
    delete:
      description: delete a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: successfully deleted product
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "400":
          description: invalid product ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: fetch detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid product ID
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: product not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: update an existing product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: update product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: successfully fetched product
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductResponse'
              type: object
        "400":
          description: invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: update a product
      tags:
      - Products
  /api/v1/products/{id}/images:
    post:
      consumes:
      - multipart/form-data
      description: upload an image for a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: successfully uploaded image
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: invalid request or file
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "403":
          description: admin access required
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Upload product Image
      tags:
      - Products
  /api/v1/search:
    get:
      description: Search Products using full text search with ranking
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Minimum price filter
        in: query
        name: min_price
        type: number
      - description: Maximum price filter
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.ProductSearchResult'
                  type: array
              type: object
        "400":
          description: Invalid Search query
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Search Products
      tags:
      - Products
  /api/v1/users/profile:
    get:
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Updated profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_dto.UserResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_deepakgudla_bookvault_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /graphql:
    post:
      consumes:
      - application/json
      description: Execute authenticated GraphQL queries and mutations. The GraphQL
        schema defines available operations.
      parameters:
      - description: GraphQL request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server.GraphQLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.GraphQLResponse'
      security:
      - BearerAuth: []
      summary: Protected GraphQL endpoint
      tags:
      - GraphQL
...
securityDefinitions:
  BearerAuth:
    description: Bearer and JWT token (bearer jwt_token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"